// bloop-utils/types/ErrorCodes.ts
export enum GenericErrorCodes {
  SOMETHING_WENT_WRONG = "GENERIC_SOMETHING_WENT_WRONG",
}

export enum TokenErrorCodes {
  INVALID_REFRESH_TOKEN = "TOKEN_INVALID_REFRESH_TOKEN",
  INVALID_ACCESS_TOKEN = "TOKEN_INVALID_ACCESS_TOKEN",
}

export enum RegisterErrorCodes {
  INVALID_USERNAME = "REGISTER_INVALID_USERNAME",
  USERNAME_TOO_SHORT = "REGISTER_USERNAME_TOO_SHORT",
  DUPLICATE_USERNAME = "REGISTER_DUPLICATE_USERNAME",
  DUPLICATE_EMAIL = "REGISTER_DUPLICATE_EMAIL",
  INVALID_EMAIL = "REGISTER_INVALID_EMAIL",
  PASSWORD_TOO_SHORT = "REGISTER_PASSWORD_TOO_SHORT",
  USERNAME_TOO_LONG = "REGISTER_USERNAME_TOO_LONG",
}

export enum LoginErrorCodes {
  INVALID_CREDENTIALS = "LOGIN_INVALID_CREDENTIALS",
}

export enum ListsErrorCodes {
  TITLE_NOT_FOUND = "LIST_TITLE_NOT_FOUND",
  TITLE_TOO_LONG = "LIST_TITLE_TOO_LONG",
  LIST_NOT_FOUND = "LIST_LIST_NOT_FOUND",
  INVALID_MEDIA = "LIST_INVALID_MEDIA",
  USERNAMES_NOT_FOUND = "LIST_USERNAMES_NOT_FOUND",
}

export enum MediaErrorCodes {
  MEDIA_NOT_FOUND = "MEDIA_MEDIA_NOT_FOUND",
}

export enum EditProfileErrorCodes {
  INVALID_USERNAME = "EDIT_PROFILE_INVALID_USERNAME",
  USERNAME_TOO_SHORT = "EDIT_PROFILE_USERNAME_TOO_SHORT",
  DUPLICATE_USERNAME = "EDIT_PROFILE_DUPLICATE_USERNAME",
  DUPLICATE_EMAIL = "EDIT_PROFILE_DUPLICATE_EMAIL",
  INVALID_EMAIL = "EDIT_PROFILE_INVALID_EMAIL",
  INVALID_PASSWORD = "EDIT_PROFILE_INVALID_PASSWORD",
  PASSWORD_TOO_SHORT = "EDIT_PROFILE_PASSWORD_TOO_SHORT",
  USERNAME_TOO_LONG = "EDIT_PROFILE_USERNAME_TOO_LONG",
}